Opensource Software/opera4.9/src/third_party/WebKit/LayoutTests/fast/spatial-navigation/snav-css-nav-direction-expected.txt

This test ensures the correctness of Spatial Navigation when CSS3 nav-up/down/left/right properties are set.

1	2	3
4	5	6
7	8	9
PASS gFocusedDocument.activeElement.getAttribute("id") is "e9"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e9"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e5"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e2"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e4"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e6"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e8"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e5"
PASS successfullyParsed is true

TEST COMPLETE
